current device id for fragments

+++ Android/app/src/main/java/com/example/myhomeapplication/Local_Persistence/Cache.java
@@ -124,7 +124,7 @@ public class Cache {
         return allMeasurements;
     }*/
 
-    public LiveData<List<Measurement>> getAllMeasurements(int deviceID, String measurementType) {
+    public LiveData<List<Measurement>> getAllMeasurements(long deviceID, String measurementType) {
         return measurementClient.getAllMeasurements(deviceID, measurementType);
     }
